requirements.txt
langchain_openai
langchain_community
langchain_core 
streamlit
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
wikipedia
arxiv
langchainhub
langchain-groq
groq
chromadb